+++ lib/go-tc/stats_summary.go
@@ -96,7 +96,7 @@ func parseTime(ts string) (time.Time, error) {
 	return time.Parse(dateFormat, ts)
 }
 
-// UnmarshalJSON Customized Marshal to force date format on statDate
+// MarshalJSON Customized Marshal to force date format on statDate
 func (ss StatsSummary) MarshalJSON() ([]byte, error) {
 	type Alias StatsSummary
 	resp := struct {
@@ -118,6 +118,7 @@ type StatsSummaryLastUpdated struct {
 	SummaryTime *time.Time `json:"summaryTime"  db:"summary_time"`
 }
 
+// MarshalJSON is a customized marshal to force date format on SummaryTime.
 func (ss StatsSummaryLastUpdated) MarshalJSON() ([]byte, error) {
 	resp := struct {
 		SummaryTime *string `json:"summaryTime"`