+++ traffic_portal/app/src/common/modules/table/userDeliveryServices/TableUserDeliveryServicesController.js
@@ -38,9 +38,6 @@ var TableUserDeliveryServicesController = function (user, deliveryServices, $con
 			"iDisplayLength": 25,
 			"aaSorting": [],
 			"columns": $scope.columns,
-			"colReorder": {
-				realtime: false
-			},
 			"initComplete": function(settings, json) {
 				try {
 					// need to create the show/hide column checkboxes and bind to the current visibility