+++ traffic_portal/app/src/common/service/application/ApplicationService.js
@@ -42,6 +42,9 @@ var ApplicationService = function($rootScope, $anchorScroll, $http, messageModel
             "buttons": [
                 { "extend": "csv", "text": "Export as CSV", "titleAttr": "Export as CSV", "className": "btn-link" }
             ],
+            "colReorder": {
+                realtime: false
+            },
             "stateSave": true,
             "scrollX": true
         });